changes as of 1435 30/5/24

+++ task4_1/Pages/DeleteAnimal.xaml.cs
@@ -31,19 +31,22 @@ public partial class DeleteAnimal : ContentPage
         {
             if (animal is Cow cow)
             {
-                ResultLabel.Text = $"Type of Animal: Cow\n" +
-                    $"Colour: {cow.Colour}\n" +
-                    $"Cost: {cow.Cost}\n" +
-                    $"Weight: {cow.Weight}\n" +
-                    $"Milk: {cow.Milk}\n";
+                ResultLabel.Text =
+                    $"Type of Animal:       Cow\n" +
+                    $"Colour:       {cow.Colour}\n" +
+                    $"Cost:     {cow.Cost}\n" +
+                    $"Weight        {cow.Weight}\n" +
+                    $"Milk:     {cow.Milk}\n";
+                    
             }
             else if (animal is Sheep sheep)
             {
-                ResultLabel.Text = $"Type of Animal: Sheep\n" +
-                    $"Colour: {sheep.Colour}\n" +
-                    $"Cost: {sheep.Cost}\n" +
-                    $"Weight: {sheep.Weight}\n" +
-                    $"Wool: {sheep.Wool}\n";
+                ResultLabel.Text =
+                    $"Type of Animal:Sheep\n" +
+                    $"Colour:{sheep.Colour}\n" +
+                    $"Cost:{sheep.Cost}\n" +
+                    $"Weight:{sheep.Weight}\n" +
+                    $"Wool:{sheep.Wool}\n";
             }
         }
         else
@@ -63,8 +66,14 @@ public partial class DeleteAnimal : ContentPage
         }
 
         Animal animal = vm.Animals.FirstOrDefault(a => a.Id == animalId);
-        vm._database.DeleteItem(animal);
-        DisplayAlert("congrats", "Successful Deletion", "OK");
-
+        if(animal == null)
+        {
+            DisplayAlert("error", "no animal found", "OK");
+        }
+        if (animal != null)
+        {
+            vm._database.DeleteItem(animal);
+            DisplayAlert("congrats", "Successful Deletion", "OK");
+        }
     }
 }
